OpenGL issues with rendering a complete cube
I've been trying really hard, but so far, I've only been able to render
the front and back faces of the wawa cube. Is it vertex data issues ?
I've been using the exact same coordinates so something must be a bit
off. Perhaps it's a texture coordinate issue? Regardless, I will try as
hard as I can in order to make a real cube.