v0.0.1-Alpha/Evolved Uni Demo Release
An amazing day for Wawacraft Evolved. This commit marks version 0.0.1
alpha of Wawacraft Evolved. Read the README.md for this version to
see some of the reasons as to what has been updated during this version.

In the upcoming version of Wawacraft:Evolved, I plan to rewrite EVERYTHING.
FROM. SCRATCH. Read the blog, which I will have updated by the time this
is uploaded online.

Much love, and
WAWACRAFT EVOLVED FOREVER!!!