Fixed up makefile
In older versions, the object files would be compiled in different
directories. I figure that it would be better to have .o files compile
over to the bin directory to keep sources separate from object
files. Makefile changes are a bit better, and I am a bit happier that
I kept my make system clean and organized now, since it was easy to
make this.