Added window class
For the rewrite, I added a better interpretation of the window
class. It is now supplied with most of the things we had before, but
in a relatively OK structure. The window class now contains methods to
rename and reassign resolutions, which is very useful.

I've also unit-tested each new function of the class, and it all works
successfully. I don't exactly know how the formal way to document unit
tests are, but I assume they're integrated into a larger commit rather
than just their own commits.

Something that might be a point of hate is that I did not use a
typedef for the function void
(*window::get_render_loop_callback(void))(window* pwin). I think
typedefs as of now wouldn't be useful, and especially not a global
typedef. So far, most things are clear enough such that a typedef
isn't exactly needed. I might consider writing a typedef for it when
it is required, but as of now, its perfectly OK.

This also brings me to talk about the structure of the program in the
rewrite. Considering how the render cycle wasn't as organized as it is
being built to be now, I should discuss how this is.

As far as I've read, I'm pretty sure self-descriptive render objects
are generally agreed on in a graphics program with the need to render
for a non-test environment. But, there will inevitably be a separation
between 2d rendering (UI, Text, images, etc.), and 3d rendering (3d
worlds, uni staring at you endlessly, wawa shrine, etc.). This means
that we will have 2 types of render objects:

2D render objects (most probably simple meshes rendered in 2d, and are
self reliant, only requiring a prepared render environment).

3D render objects (complex meshes, with independent shaders, and
dependent on a greater set of information)

In the old code structure, for 3d objects, I assigned 3d objects to be
rendered through a camera class, which would supply the view and
projection matrices to complete rendering in the shader. I don't
exactly remember how the shaders were dealt with, but IIRC the shaders
were independently sourced by caller. I made the bad mistake of making
a 3d object also double as a 2d object, with heavy responsibilities on
the caller to structure vertex information and supply textures.

Textures also are a bit of a hard thing to understand. I know that
graphics programs usually try to sparingly load textures, and keep
them in easy access for practically everybody. Would that mean that
textures are not associated with meshes? I mean UV coordinates are
associated with the mesh, so I see it optimal to associate
textures(even if a copy) with the mesh as a fallback. Considering it
as an engine, it's better to assign textures to objects rather than
keep them in a public domain array where any mesh can easily source
textures.

But for 3d objects, a 3d environment must be established on top of the
existing 2d environment. This environment should be able to be
separated from the 2d environment, so that they work simultaneously. I
assume writing 3d objects to the screen before writing 2d objects
would work. But this idea of setting up environments before rendering
is a good idea.

We could make a bool or a public flag that both classes can access, or
have them exchange flags from each other. static bool 3dengaged and
static bool 2dengaged would work. Maybe itd be a better structural
choice to have the caller take responsibility for which render mode is
engaged.

So we understand now that unifying the render cycle, we must create a
render environment object which manages both 3d and 2d rendering. It
is a factor required in the independent rendering of 3d and 2d
objects. They will have the camera objects pointer in them, and will
allow the caller to immediately access the camera in use. And 2d
objects will simply not use the 3d tools.

And for control environments... That requires for me to write
keymapping functions to make a virtual keyboard object, and assign
functions to every single key

More realistically, we could define key objects, which handle the
reception of different keys, and allow the caller to get their
state. Then I could make getKeyStatus(KEY_ENUM)==true turn into an
operator overload like KEY_OBJECT~ which returns the state of the key.

Regardless, this is a good way to start development for the
rewrite. I've got a mount of studying outside of this to do, so I'll
make a rewrite for the render environment when I have actual free time
to spare.

WAWACRAFT:EVOLVED FOREVER!!!!