Added render_environment class
In a flash of genius, I decided that a render environment should be
made. It will contain the necessary tools and supplies needed to generate
a 3D image. So far, it is simple.

I also made sure to add more error codes for when things mess up, both the
window and render_environment class have function pointers, which means
you HAVE to keep pointer safety.

I unit-tested the feature of setting the colors, but I will not write
unit tests to see if I can properly change things about the render
environment because I already know the function works.

WAWACRAFT:EVOLVED FOREVER!!!!!