Finally set up a basic system
So my ideas for the rendering environment code did work! I just had to change up a
few things in there to ensure good codebase management. The render environment will
allow you to mess around with shaders and meshes in the render callback instead of
having an empty shell of a render callback as was happening in the previous source.

Since the render environment is just a beautiful little shell for rendering AROUND
mainly what is the window class, its render callback function is simply a frontend
for calling window::run_render_loop() (now with void pointers for the shader and mesh).
This does pose a few risks of "What if we mess something up during casting process?"
but I made the design choice that all of this will be documented in the docs and that
the risks are only present if I forget to patch around the security risks in the
render function.

Regardless, I did set up some basic functions for meshes, and meshes will be part of
a large family tree of meshes all stemming from mesh_base. Extra functionality is added
that way in order to give anybody using the engine a choice to interact with a type
of mesh from the class family depending on what functionality they would like. UI
meshes might have completely different systematic frontends than 3d object meshes, and
meshes for the skybox could also be completely different, so on.

All is well and most of the new code actually works. I want to at least finish semi
OK development before New Years, and more importantly, before Jan 19th. Low-level things
mostly dealt with, future development will entail setting up a handwritten math
library that actually works and a proper keyboard input control system with
keymapping to allow later on control edits.

Well, that's a nice commit out of the way, now to go play Extreme Tux Racer.