Adding texture support
After a bit of unit testing, it is quite obvious that if you are going to play with the
mesh display attributes, you better be prepared to edit the shaders that you use to
reflect that choice. Regardless, it wasn't very hard setting up textures in here, so I
guess this will bring us into making the math library to prepare for our 3d projection
and the MVP matrix trilogy.

Also, Happy New Years! This is the first commit of the year to Wawacraft Evolved!

Better get done with a proper demo before Jan 19th.

WAWACRAFT:EVOLVED FOREVER!!!!