Changes to the structure of render_environment and its render callback
As it is already known, render_environment is a render environment
holding a pointer to both types of meshes, which casts these types
to void pointers to get them to pass through to the class window, which
later sends it to a user-defined callback which can cast these back
in a canonical use of the engine to a mesh_3d or whatever they may so
like :).

Now these pointers have been updated to mesh_base** and mesh_3d** to allow
for more meshes to be given, preferrably through a mesh*[] which holds pointers
to all the meshes you need to render.

I've seen relatively good success in fixing this issue on the rewrite,
and will try to see what I can do to try and make textures work
such that multiple textures can render at once. Currently, there is a
bug where only the first binded texture works. Perhaps you need to bind
the textures manually first? Though I do do that already...

Well, that's a good commit for the day, isn't it?

WAWACRAFT:EVOLVED FOREVER!!!