SPRITESHEET SUPPORT
So we now have working spritesheet support on chunks. The demo
provided shows a chunk with a number of different textures and
an amazing view. We now have chunks that can theoretically have
any shape and any texture from a range of 4096 different
textures in one chunk.

This does mean that I will be removing the old textures from
the repo. Say goodbye shaven wawa and unkept uni, you were far
too shiny and 4k for this world. And also the separate grass
texture as it is not needed anymore, as well as other, useless
test files from the older project.