Added shading
I added shading to the nocol shader, as well as
providing shading options for each vertex in
mesh_base.

Secondly, there is also a bug fixed in mesh
generation for chunks wherein it would not
render all the right faces correctly because
I incremented the actual value instead of manually
doing it. To save myself the headache of trying to
find out how to increment as a value and not
write it to the iterators, I just put +1 and -1
in there instead. My code was already perfect but
I guess it didnt work that way then.