Fixed the bug and realized my thing is like super fast
It wasn't an issue of rounding error, in fact it turns out I have
a very intelligent design for it. The error was that I was unloading
chunks AFTER I created the visible list(which holds references to
indexes in the chunk array), which invalidated the visibility list in
some cases. Now terrain works completely with all kinds of coordinates,
as I unload unneeded chunks BEFORE I find the visible chunks and
set their references for the frame.